Rewrite trace names if they're from HTTPoison
Look for URLs in the first or second arg of a traced external call, and
if present, use its hostname as the name of the trace.

This is required for back compatibility with TRRs old NewRelic agent.
Once the mainline branch supports this type of behavior, this code can
be removed